feat(web): split account dashboard routes
Move account-owned billing, machine, and certificate surfaces under /account while keeping dashboard shims for existing entry points.

Extract the coupling forensics panel, add the forensics OAuth scope contract, and update API, bot, web, and route tests for the new surfaces.

Resolve the OpenTelemetry core audit advisory by aligning shared observability dependencies with the v2 provider API.
