+++ apps/web/src/routes/_authenticated/account/machines.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from '@tanstack/react-router';
+
+export const Route = createFileRoute('/_authenticated/account/machines')({});
