+++ apps/web/src/routes/_authenticated/dashboard/billing.tsx
@@ -1,4 +1,4 @@
-import { createFileRoute } from '@tanstack/react-router';
+import { createFileRoute, redirect } from '@tanstack/react-router';
 
 interface DashboardBillingSearch {
   plan?: string;
@@ -7,6 +7,8 @@ interface DashboardBillingSearch {
   source?: string;
 }
 
+// Billing now lives in the account area. Keep this path as a redirect so existing
+// deep links (Polar handoff, Unity, the API connect flow) keep working.
 export const Route = createFileRoute('/_authenticated/dashboard/billing')({
   validateSearch: (search: Record<string, unknown>): DashboardBillingSearch => ({
     plan: typeof search.plan === 'string' ? search.plan : undefined,
@@ -14,4 +16,7 @@ export const Route = createFileRoute('/_authenticated/dashboard/billing')({
     portal: typeof search.portal === 'string' ? search.portal : undefined,
     source: typeof search.source === 'string' ? search.source : undefined,
   }),
+  beforeLoad: ({ search }) => {
+    throw redirect({ to: '/account/billing', search, replace: true });
+  },
 });
