+++ apps/web/src/routes/_authenticated/dashboard/certificates.tsx
@@ -1,3 +1,9 @@
-import { createFileRoute } from '@tanstack/react-router';
+import { createFileRoute, redirect } from '@tanstack/react-router';
 
-export const Route = createFileRoute('/_authenticated/dashboard/certificates')({});
+// The standalone certificates page was retired. Authorized machines now live in the
+// account area; keep this path as a redirect for any lingering deep links.
+export const Route = createFileRoute('/_authenticated/dashboard/certificates')({
+  beforeLoad: () => {
+    throw redirect({ to: '/account/machines', replace: true });
+  },
+});
