+++ apps/web/src/routes/_authenticated/dashboard/forensics.tsx
@@ -1,3 +1,9 @@
-import { createFileRoute } from '@tanstack/react-router';
+import { createFileRoute, redirect } from '@tanstack/react-router';
 
-export const Route = createFileRoute('/_authenticated/dashboard/forensics')({});
+// Coupling forensics is now a view inside the private VPM package workspace.
+// Keep this path as a redirect for the Discord bot link and any existing deep links.
+export const Route = createFileRoute('/_authenticated/dashboard/forensics')({
+  beforeLoad: () => {
+    throw redirect({ to: '/dashboard/packages', search: { view: 'forensics' }, replace: true });
+  },
+});
