+++ apps/web/src/routes/_authenticated/dashboard/packages.tsx
@@ -1,3 +1,11 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-export const Route = createFileRoute('/_authenticated/dashboard/packages')({});
+interface DashboardPackagesSearch {
+  view?: 'registry' | 'forensics';
+}
+
+export const Route = createFileRoute('/_authenticated/dashboard/packages')({
+  validateSearch: (search: Record<string, unknown>): DashboardPackagesSearch => ({
+    view: search.view === 'forensics' ? 'forensics' : undefined,
+  }),
+});
