+++ package.json
@@ -8,6 +8,7 @@
   ],
   "overrides": {
     "@opentelemetry/auto-instrumentations-node": ">=0.75.0",
+    "@opentelemetry/core": "2.8.0",
     "@opentelemetry/exporter-prometheus": ">=0.217.0",
     "@opentelemetry/sdk-node": ">=0.217.0",
     "@babel/core": "7.29.7",
