+++ packages/shared/package.json
@@ -22,11 +22,11 @@
   "dependencies": {
     "@infisical/sdk": "^4.0.6",
     "@opentelemetry/api": "^1.9.1",
-    "@opentelemetry/context-async-hooks": "^1.30.1",
-    "@opentelemetry/core": "^1.30.1",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.57.2",
-    "@opentelemetry/resources": "^1.30.1",
-    "@opentelemetry/sdk-trace-base": "^1.30.1",
+    "@opentelemetry/context-async-hooks": "^2.8.0",
+    "@opentelemetry/core": "^2.8.0",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.219.0",
+    "@opentelemetry/resources": "^2.8.0",
+    "@opentelemetry/sdk-trace-base": "^2.8.0",
     "@tempojs/common": "^0.0.12"
   },
   "devDependencies": {
